requirements.txt
web3>=6.11.0
eth-account>=0.10.0
eth-abi>=5.0.0
requests>=2.31.0
python-dotenv>=1.0.0
tenacity>=8.2.0
sqlalchemy>=2.0.0
uvicorn>=0.30.0
fastapi>=0.110.0
rich>=13.7.0